iswellformed => !ismalformed; make test stricter
despite generally trying to avoid double negatives, it seems that
"malformed" is the clearer and more common concept to test for here.
it was also a bit confusing to have `MalformedCharError` but then
have `iswellformed` as the predicate instead of `ismalformed`.

testing for well-formedness more strictly doesn't seem to add too
much complexity here; if performance becomes an issue, we'll need a
`@inbounds`-like annotation anyway.